common: employ a globally available convert_context
Everytime you call convert_to<>(...), an iconv context is created and
soon destroyed. On glibc-2.32's iconv, each call of the {iconv_open,
iconv_close} pair measures in at about 283 ns, whereas a state reset
of an existing iconv context measures just 33 ns. However, there is
already an implicit state reset issued by doconvert(), so it counts
as 0 ns.

Zarafa knew this too, which is why the class convert_context exists,
but it is not used everywhere, because it was a lot to type everytime
(and this becomes evident with the removal of its callsites in the
next commits). C++11 offers the necessary features that we can now
make a global drop-in. The cost is 3 ns. (94x improvement over
before)