common: redo persistence for convert_context
As a result of dlclose, the pointer obtained by a previous
typeid(x).name() can become invalid. This could likewise happen for
string literal pointers whose origin is through a dlopened library
set. The C++ and POSIX standards say nothing about requirements, so
assume the worst and always persist code and type strings to the full
extent. This would previously manifest in a crash when calling
convert_context functions anytime after libkcclient.so was closed
(usually from the time of ~M4LMsgServiceAdmin onwards).