icalmapi: emit allday events as TZ-ful 24h events
Cease emitting floating dates/floating timepoints. These are a really
bad idea, because they have arbitrary meaning due to lack of a
timezone.

When the iCal input is converted to a PT_SYSTIME (struct FILETIME),
we *must* be aware of the zone, because the PT_SYSTIME MAPI type
demands to be in UTC. (MS-OXOCAL 3.1.5.5 even acknowledges this.)

References: KC-1664